refactor formatCreator and formatName into utility functions. update imports.
