+++ frontend/src/utilities/types.ts
@@ -1,0 +1,3 @@
+import { Geojson_Checksum_NormalizedCreator } from "../graphql/types";
+
+export type CreatorName = Pick<Geojson_Checksum_NormalizedCreator, "name">
